refactor: move files to common directory
