--- src/components/UI/index.ts
@@ -1,7 +1,0 @@
-import MyButton from "@/components/UI/MyButton.vue";
-import MySelect from "@/components/UI/MySelect.vue";
-
-export default [
-  MyButton,
-  MySelect
-]
