+++ src/components/common/index.ts
@@ -1,4 +1,6 @@
 import FadeTransition from './FadeTransition.vue';
 import AppSpinner from './AppSpinner.vue';
+import MyButton from './MyButton.vue';
+import MySelect from './MySelect.vue';
 
-export { FadeTransition, AppSpinner };
+export default [FadeTransition, AppSpinner, MyButton, MySelect];
